Guardado de venta
Se complementa el guardado de venta

+++ SistemaFarmacia.Entidades/Negocio/VentaDetalle.cs
@@ -12,7 +12,7 @@ namespace SistemaFarmacia.Entidades.Negocio
         public Int64 IdVentaDetalle { get; set; }
         public int IdProducto { get; set; }
         public string ClaveProducto { get; set; }
-        public Int16 Cantidad { get; set; }
+        public decimal Cantidad { get; set; }
         public decimal Precio { get; set; }
         public string Descripcion { get; set; }
         public decimal Total { get; set; }
